updated tests, making them truly reliable
