+++ package.json
@@ -50,6 +50,7 @@
   },
   "devDependencies": {
     "babel-preset-env": "^1.7.0",
-    "babel-register": "^6.26.0"
+    "babel-register": "^6.26.0",
+    "sequelize-test-helpers": "^1.1.0"
   }
 }
