+++ tests/api.spec.js
@@ -1,22 +1,22 @@
-let request = require('supertest');
 const api = require('../routes');
-//const app = require('../server');
+const app = require('../app');
+let request = require('supertest');
 
 // TESTS THE API
-describe('Testing the API', ()=>{
-
+describe('Test the API', ()=>{
   test('API Router should be defined', () => {
     expect(api).toBeDefined();
   });
   test('API Router should be an object', () => {
     expect(typeof api).toBe("object");
   });
-
-/*
-  test('API root path should be accessible', async (done) => {
-    request = request('http://localhost:3000/api');
-    await request.get('/').expect(200,done);
-    done();
+  test('API root path should return 200', () => {
+    return request(app).get('/').expect(200);
   });
-*/
-})
+  test('API /api endpoint should return 200', () => {
+    return request(app).get('/api').expect(200);
+  });
+  /*test("API /api/auth should return 401 'Unauthorized'", () => {
+    return request(app).get('/api/auth').expect(401);
+  });*/
+});
